==> scripts/xx_Archive/Toy_Model/1_ToyModel.Rmd <==
---
title: "Toy Model"
author: "Julia Lawson, UCSB"
date: "2023-09-04"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(here)
library(knitr)
library(stargazer)
library(reshape2)
```

## 0. Initial Parameters

```{r}

# Growth Parameters

x_a = 70
h_a = 0
k_a = 100
r_a = 0.6

x_b = 20
h_b = 0
k_b = 100
r_b = 0.7

# Movement Parameters
M_ab = 0.7 # Adult movement A -> B 
M_ba = 0.6 # Adult movement B -> A
M_aa = 1 - M_ab
M_bb = 1 - M_ba
D_ab = 0.2 # Larval dispersal A -> B
D_ba = 0.3 # Larval dispersal B -> A
D_aa = 1 - D_ab
D_bb = 1 - D_ba

# Survival Parameters
mu_a = 0.7
mu_b = 0.8
sigma_a = 0.9
sigma_b = 0.6

```

## I. Adult Movement

```{r}

# Escapement
e_a = x_a - h_a
e_b = x_b - h_b

```

## II. Larval Production

```{r}

# Patch-specific population growth Patch A
growth_a = function(e_a, k.=k_a, r.=r_a){
  y_a = r.*e_a*(1-e_a/k.)
  return(y_a)
}

# Patch-specific population growth Patch B
growth_b = function(e_b, k.=k_b, r.=r_b){
  y_b = r.*e_b*(1-e_b/k.)
  return(y_b)
}

```

```{r}

result1 = data.frame(e_a = seq(1,k_a), e_b = seq(1,k_b)) %>%
  mutate(y_a = growth_a(e_a)) %>% 
  mutate(y_b = growth_b(e_b))

```

### III. Larval contribution at t+1 including survival

```{r}

new_a = function(y_b = y_b, y_a = y_a, sigma_b.=sigma_b,D_bb. = D_bb, D_ab. = D_ab){
  a_1_larval =  (sigma_a*(y_a*D_aa + y_b*D_ba))
  return(a_1_larval)
}

new_b = function(y_a = y_a, y_b = y_b, sigma_a.=sigma_a, D_aa. = D_aa, D_ba. = D_ba){
  b_1_larval =  (sigma_b*(y_b*D_bb + y_a*D_ab))
  return(b_1_larval)
}


result1 = result1 %>% 
  mutate(a_1_larval = new_a(y_b, y_a)) %>% 
  mutate(b_1_larval = new_b(y_a, y_b))

```

### IV. Adult contribution at t+1 including survival

```{r}

adult_a = function(e_b = e_b, e_a = e_a, mu_a.=mu_a, mu_b.=mu_b, M_bb. = M_bb, M_ab. = M_ab){
  a_1_adult = (mu_a*(e_a*M_aa) + mu_b*(e_b*M_ba))
  return(a_1_adult)
}

adult_b = function(e_a = e_a, e_b = e_b, mu_a.=mu_a, mu_b.=mu_b, M_aa. = M_aa, M_ba. = M_ba){
  b_1_adult = (mu_b*(e_b*M_bb) + mu_a*(e_a*M_ab))
  return(b_1_adult)
}

result1 = result1 %>% 
  mutate(a_1_adult = adult_a(e_b, e_a)) %>% 
  mutate(b_1_adult = adult_b(e_a, e_b))

```

### V. Full population in next time step

```{r}

result1 = result1 %>% 
  mutate(a_t1 = a_1_larval + a_1_adult) %>% 
  mutate(b_t1 = b_1_larval + b_1_adult)

```


```{r}

# Figure of Steady States

for_graph = result1

plot_a = for_graph %>%
  ggplot(aes(x = e_a, 
             y = a_t1)) +
  theme_bw() +
  geom_line(linewidth = 0.7) +
  geom_abline(slope=1, linetype="dashed", color="red") #+ # 45 degree line to find steady state
  #geom_vline(aes(xintercept=x_a), linetype="dashed") # initial population, x_a

plot_b = for_graph %>%
  ggplot(aes(x = e_b, 
             y = b_t1)) +
  theme_bw() +
  geom_line(linewidth = 0.7) +
  geom_abline(slope=1, linetype="dashed", color="red") #+ # 45 degree line to find steady state
  #geom_vline(aes(xintercept=x_b), linetype="dashed") #initial population, x_b

plot_a
plot_b
```

```{r}

# Table of Steady States

result1 = result1 %>% 
  mutate(diff_b = b_t1-e_b) %>% 
  mutate(diff_a = a_t1-e_a)

steady_states = c()
  
for(variable in c('diff_a', 'diff_b')){
  # Find where difference equations change sign
  ss = which(sign(eval(parse(text = paste0('result1$',variable)))) != lag(sign(eval(parse(text = paste0('result1$', variable))))))
  if(length(ss)==1){
      steady_states = append(steady_states, c(result1[ss, 1]))
  }
}

final_ = as.data.frame(steady_states) %>% 
  add_column(patch = c("A","B")) %>% 
  select(patch,steady_states)
final_
```

### Economic Model

```{r}

p_a = 40
c_a = 20
p_b = 30
c_b = 10

# discount factor
delta = 0.95

# Calculate marginal profit
b_a = p_a - c_a
b_b = p_b - c_b

```

### I. Sole Owner Solution

```{r}

# Calculate phi_a and phi_b
phi_a <- (b_a/delta - b_a*mu_a*M_aa - b_b*mu_a*M_ab) / 
                 (b_a*sigma_a*D_aa + b_b*sigma_b*D_ab)

phi_b <- (b_b/delta - b_b*mu_b*M_bb - b_a*mu_b*M_ba) / 
                 (b_b*sigma_b*D_bb + b_a*sigma_a*D_ba)

# Calculate e_star_a and e_star_b
e_star_a <- (k_a * (r_a - phi_a)) / (2 * r_a)
e_star_b <- (k_b * (r_b - phi_b)) / (2 * r_b)

final_ = mutate(final_, e_star = c(e_star_a, e_star_b))

```

### II. Non-cooperative Solution

```{r}

phi_hat_a <- (1/delta - mu_a*M_aa) / (sigma_a*D_aa)
phi_hat_b <- (1/delta - mu_b*M_bb) / (sigma_b*D_bb)

e_hat_a <- (k_a * (r_a - phi_hat_a)) / (2 * r_a)
e_hat_b <- (k_b * (r_b - phi_hat_b)) / (2 * r_b)

final_ = mutate(final_, e_hat = c(e_hat_a, e_hat_b))

```

### III. Plots with sole owner solution (e_star) and non-cooperative solution (e_hat)

```{r}

plot_a + geom_vline(aes(xintercept=e_star_a), color="purple", linetype="dashed") +
  geom_vline(aes(xintercept=e_hat_a), color="green", linetype="dashed")

plot_b + geom_vline(aes(xintercept=e_star_b), color="purple", linetype="dashed") +
  geom_vline(aes(xintercept=e_hat_b), color="green", linetype="dashed")

```
